Create common components to upload images
- Remove un-necesary components
- Redirect to edit page for all the major forms
