+++ src/components/domain/ImageWithCaptionInput/styles.module.css
