+++ src/views/DrefApplicationForm/Overview/index.tsx
@@ -36,6 +36,7 @@ import DistrictSearchMultiSelectInput, {
     type DistrictItem,
 } from '#components/domain/DistrictSearchMultiSelectInput';
 import useDisasterType from '#hooks/domain/useDisasterType';
+import ImageWithCaptionInput from '#components/domain/ImageWithCaptionInput';
 
 import {
     DISASTER_FIRE,
@@ -45,7 +46,6 @@ import {
     TYPE_LOAN,
 } from '../common';
 import { type PartialDref } from '../schema';
-import ImageWithCaptionInput from './ImageWithCaptionInput';
 import CopyFieldReportSection from './CopyFieldReportSection';
 import styles from './styles.module.css';
 import i18n from './i18n.json';
@@ -446,6 +446,7 @@ function Overview(props: Props) {
                 >
                     <ImageWithCaptionInput
                         name="event_map_file"
+                        url="/api/v2/dref-files/"
                         value={value?.event_map_file}
                         onChange={setFieldValue}
                         error={getErrorObject(error?.event_map_file)}
@@ -465,6 +466,7 @@ function Overview(props: Props) {
                 >
                     <ImageWithCaptionInput
                         name="cover_image_file"
+                        url="/api/v2/dref-files/"
                         value={value?.cover_image_file}
                         onChange={setFieldValue}
                         error={getErrorObject(error?.cover_image_file)}
