--- src/views/DrefFinalReportForm/EventDetail/MultiImageWithCaptionInput/styles.module.css
@@ -1,26 +1,0 @@
-.multi-image-with-caption-input {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: var(--go-ui-spacing-md);
-
-    .file-list {
-        display: grid;
-        grid-gap: var(--go-ui-spacing-md);
-        grid-template-columns: repeat(auto-fit, minmax(16rem, 1fr));
-
-        .preview-and-caption {
-            display: flex;
-            flex-direction: column;
-            gap: var(--go-ui-spacing-xs);
-
-            .preview {
-                background-color: var(--go-ui-color-background);
-                width: 100%;
-                height: 12rem;
-                object-fit: contain;
-                object-position: center center;
-            }
-        }
-    }
-}
