+++ src/views/DrefOperationalUpdateForm/EventDetail/index.tsx
@@ -10,6 +10,7 @@ import BooleanInput from '#components/BooleanInput';
 import TextArea from '#components/TextArea';
 import DateInput from '#components/DateInput';
 import useTranslation from '#hooks/useTranslation';
+import MultiImageWithCaptionInput from '#components/domain/MultiImageWithCaptionInput';
 
 import {
     TYPE_IMMINENT,
@@ -19,7 +20,6 @@ import {
 } from '../common';
 import { type PartialOpsUpdate } from '../schema';
 
-import MultiImageWithCaptionInput from './MultiImageWithCaptionInput';
 import i18n from './i18n.json';
 import styles from './styles.module.css';
 
@@ -233,6 +233,7 @@ function EventDetail(props: Props) {
                         <MultiImageWithCaptionInput
                             // FIXME: use translation
                             label="Select images"
+                            url="/api/v2/dref-files/multiple/"
                             name="images_file"
                             value={value.images_file}
                             onChange={setFieldValue}
