--- src/views/DrefOperationalUpdateForm/Overview/ImageWithCaptionInput/styles.module.css
@@ -1,11 +1,0 @@
-.image-with-caption-input {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    gap: var(--go-ui-spacing-md);
-
-    .preview {
-        width: 10rem;
-        height: 10rem;
-    }
-}
