+++ src/views/DrefOperationalUpdateForm/Overview/index.tsx
@@ -37,6 +37,7 @@ import DistrictSearchMultiSelectInput, {
     type DistrictItem,
 } from '#components/domain/DistrictSearchMultiSelectInput';
 import useDisasterType from '#hooks/domain/useDisasterType';
+import ImageWithCaptionInput from '#components/domain/ImageWithCaptionInput';
 
 import {
     DISASTER_FIRE,
@@ -47,7 +48,6 @@ import {
     ONSET_SUDDEN,
 } from '../common';
 import { type PartialOpsUpdate } from '../schema';
-import ImageWithCaptionInput from './ImageWithCaptionInput';
 import styles from './styles.module.css';
 import i18n from './i18n.json';
 
@@ -489,6 +489,7 @@ function Overview(props: Props) {
                 >
                     <ImageWithCaptionInput
                         name="event_map_file"
+                        url="/api/v2/dref-files/"
                         value={value?.event_map_file}
                         onChange={setFieldValue}
                         error={getErrorObject(error?.event_map_file)}
@@ -508,6 +509,7 @@ function Overview(props: Props) {
                 >
                     <ImageWithCaptionInput
                         name="cover_image_file"
+                        url="/api/v2/dref-files/"
                         value={value?.cover_image_file}
                         onChange={setFieldValue}
                         error={getErrorObject(error?.cover_image_file)}
